made component more compat
